refactor: reader -> adpater 클래스명 수정
