+++ src/main/java/codesquad/adapter/StaticFileAdapterSpec.java
@@ -1,8 +1,8 @@
-package codesquad.reader;
+package codesquad.adapter;
 
 import java.io.IOException;
 
-public interface StaticFileReaderSpec {
+public interface StaticFileAdapterSpec {
     public String readFileLines(String path) throws IOException;
 
 }
